requirements.txt
Django==1.7.1
psycopg2==2.5.4
django-taggit==0.12.2